Evolucao do algoritmo LLL

+++ Mathematics/AlgebraicStructures/Implementations/DoubleNearestInteger.cs
@@ -8,16 +8,16 @@
     /// <summary>
     /// Permite obter o inteiro mais próximo de um número decimal.
     /// </summary>
-    public class DoubleNearestInteger : INearest<double, int>
+    public class DoubleNearestInteger : INearest<double, double>
     {
         /// <summary>
         /// Obtém o inteiro que mais se aproxima do número especificado.
         /// </summary>
         /// <param name="source">O número.</param>
         /// <returns>O inteiro.</returns>
-        public int GetNearest(double source)
+        public double GetNearest(double source)
         {
-            return (int)Math.Round(source);
+            return Math.Round(source);
         }
     }
 }
